+++ src/main/java/calculator/io/InputProcessor.java
@@ -1,5 +1,6 @@
-package calculator;
+package calculator.io;
 
+import calculator.utils.ExceptionHandler;
 import camp.nextstep.edu.missionutils.Console;
 import java.util.ArrayList;
 import java.util.Arrays;
