+++ src/main/java/calculator/utils/Calculator.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.utils;
 
 import java.util.List;
