+++ src/test/java/calculator/ExceptionHandlerTest.java
@@ -2,6 +2,7 @@ package calculator;
 
 import static org.assertj.core.api.Assertions.*;
 
+import calculator.utils.ExceptionHandler;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Stream;
